# doc/specs/openapi.yaml
openapi: 3.1.0
info:
  title: DashFlow API
  version: 0.2.0
  description: |
    REST contract for DashFlow MVP — workspaces, projects, tasks, comments, JWT access + httpOnly refresh cookie.
servers:
  - url: http://localhost:8000/api/v1
    description: Local
security:
  - bearerAuth: []
paths:
  /health:
    get:
      summary: Health check
      operationId: healthCheck
      security: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    example: ok
  /auth/register:
    post:
      summary: Register
      operationId: authRegister
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "409":
          description: Email already registered
  /auth/login:
    post:
      summary: Login
      operationId: authLogin
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: OK — sets httpOnly refresh cookie (Path `/api/v1/auth`)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "401":
          description: Invalid credentials
  /auth/refresh:
    post:
      summary: Refresh access token
      operationId: authRefresh
      security: []
      description: Reads refresh JWT from cookie; returns new access token.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefreshResponse"
        "401":
          description: Missing or invalid refresh cookie
  /auth/logout:
    post:
      summary: Clear refresh cookie
      operationId: authLogout
      security: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
  /me:
    get:
      summary: Current user
      operationId: meGet
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserRead"
        "401":
          description: Not authenticated
  /workspaces:
    get:
      summary: List workspaces for current user
      operationId: workspacesList
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkspaceRead"
    post:
      summary: Create workspace (creator becomes owner)
      operationId: workspacesCreate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkspaceCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceRead"
  /workspaces/{workspace_id}:
    get:
      summary: Get workspace
      operationId: workspacesGet
      parameters:
        - $ref: "#/components/parameters/workspaceId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceRead"
        "403":
          description: Not a member
        "404":
          description: Not found
  /workspaces/{workspace_id}/projects:
    get:
      summary: List projects in workspace
      operationId: projectsList
      parameters:
        - $ref: "#/components/parameters/workspaceId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectRead"
    post:
      summary: Create project
      operationId: projectsCreate
      parameters:
        - $ref: "#/components/parameters/workspaceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectRead"
  /projects/{project_id}:
    get:
      summary: Get project
      operationId: projectsGet
      parameters:
        - $ref: "#/components/parameters/projectId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectRead"
    patch:
      summary: Update project
      operationId: projectsPatch
      parameters:
        - $ref: "#/components/parameters/projectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectRead"
    delete:
      summary: Delete project
      operationId: projectsDelete
      parameters:
        - $ref: "#/components/parameters/projectId"
      responses:
        "204":
          description: Deleted
  /projects/{project_id}/tasks:
    get:
      summary: List tasks in project
      operationId: tasksList
      parameters:
        - $ref: "#/components/parameters/projectId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TaskRead"
    post:
      summary: Create task
      operationId: tasksCreate
      parameters:
        - $ref: "#/components/parameters/projectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskRead"
  /tasks/{task_id}:
    get:
      summary: Get task
      operationId: tasksGet
      parameters:
        - $ref: "#/components/parameters/taskId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskRead"
    patch:
      summary: Update task
      operationId: tasksPatch
      parameters:
        - $ref: "#/components/parameters/taskId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskRead"
    delete:
      summary: Delete task
      operationId: tasksDelete
      parameters:
        - $ref: "#/components/parameters/taskId"
      responses:
        "204":
          description: Deleted
  /tasks/{task_id}/comments:
    get:
      summary: List comments on task
      operationId: commentsList
      parameters:
        - $ref: "#/components/parameters/taskId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CommentRead"
    post:
      summary: Add comment
      operationId: commentsCreate
      parameters:
        - $ref: "#/components/parameters/taskId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommentRead"
  /comments/{comment_id}:
    patch:
      summary: Update own comment
      operationId: commentsPatch
      parameters:
        - $ref: "#/components/parameters/commentId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommentRead"
    delete:
      summary: Delete own comment
      operationId: commentsDelete
      parameters:
        - $ref: "#/components/parameters/commentId"
      responses:
        "204":
          description: Deleted
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    workspaceId:
      name: workspace_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    projectId:
      name: project_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    taskId:
      name: task_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    commentId:
      name: comment_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
    UserRead:
      type: object
      required: [id, email, name]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
          maxLength: 128
        name:
          type: string
          default: ""
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    TokenResponse:
      type: object
      required: [access_token, token_type, user]
      properties:
        access_token:
          type: string
        token_type:
          type: string
          default: bearer
        user:
          $ref: "#/components/schemas/UserRead"
    RefreshResponse:
      type: object
      required: [access_token, token_type]
      properties:
        access_token:
          type: string
        token_type:
          type: string
          default: bearer
    WorkspaceCreate:
      type: object
      required: [name]
      properties:
        name:
          type: string
          maxLength: 255
        slug:
          type: string
          maxLength: 255
          nullable: true
    WorkspaceRead:
      type: object
      required: [id, name, slug, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ProjectCreate:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        color:
          type: string
          nullable: true
        sort_order:
          type: integer
          default: 0
    ProjectUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        color:
          type: string
          nullable: true
        sort_order:
          type: integer
    ProjectRead:
      type: object
      required: [id, workspace_id, name, description, color, sort_order, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        workspace_id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        color:
          type: string
          nullable: true
        sort_order:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    TaskStatus:
      type: string
      enum:
        - todo
        - in_progress
        - in_review
        - done
        - blocked
        - cancelled
    TaskPriority:
      type: string
      enum:
        - none
        - low
        - medium
        - high
        - urgent
    TaskCreate:
      type: object
      required: [title]
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/TaskStatus"
        priority:
          $ref: "#/components/schemas/TaskPriority"
        due_date:
          type: string
          format: date
          nullable: true
        start_date:
          type: string
          format: date
          nullable: true
        assignee_id:
          type: string
          format: uuid
          nullable: true
        position:
          type: integer
          default: 0
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/TaskStatus"
        priority:
          $ref: "#/components/schemas/TaskPriority"
        due_date:
          type: string
          format: date
          nullable: true
        start_date:
          type: string
          format: date
          nullable: true
        assignee_id:
          type: string
          format: uuid
          nullable: true
        position:
          type: integer
    TaskRead:
      type: object
      required:
        - id
        - project_id
        - title
        - status
        - priority
        - position
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
        priority:
          type: string
        due_date:
          type: string
          format: date
          nullable: true
        start_date:
          type: string
          format: date
          nullable: true
        assignee_id:
          type: string
          format: uuid
          nullable: true
        created_by_id:
          type: string
          format: uuid
          nullable: true
        position:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CommentCreate:
      type: object
      required: [body]
      properties:
        body:
          type: string
          minLength: 1
          maxLength: 20000
    CommentUpdate:
      type: object
      required: [body]
      properties:
        body:
          type: string
    CommentRead:
      type: object
      required: [id, task_id, user_id, body, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        task_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        body:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time